+++ src/StandardIteratorInterface.php
@@ -18,6 +18,7 @@
 
 namespace CloudCreativity\Utils\Collection;
 
+use Closure;
 use Countable;
 use IteratorAggregate;
 use OutOfBoundsException;
@@ -28,7 +29,12 @@ interface StandardIteratorInterface extends IteratorAggregate, Countable
     /**
      * @return array
      */
-    public function getAll();
+    public function all();
+
+    /**
+     * @return Collection
+     */
+    public function collect();
 
     /**
      * @return mixed
@@ -45,28 +51,28 @@ interface StandardIteratorInterface extends IteratorAggregate, Countable
     public function last();
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return StandardIteratorInterface
      */
-    public function filter(callable $callback);
+    public function filter(Closure $callback);
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return StandardIteratorInterface
      */
-    public function reject(callable $callback);
+    public function reject(Closure $callback);
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return bool
      */
-    public function every(callable $callback);
+    public function every(Closure $callback);
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return bool
      */
-    public function any(callable $callback);
+    public function any(Closure $callback);
 
     /**
      * @return bool