+++ src/StandardIteratorTrait.php
@@ -18,22 +18,31 @@
 
 namespace CloudCreativity\Utils\Collection;
 
+use Closure;
 use OutOfBoundsException;
 
 trait StandardIteratorTrait
 {
 
     /**
-     * @var Collection|null
+     * @var Collection
      */
     protected $stack;
 
     /**
      * @return array
      */
-    public function getAll()
+    public function all()
     {
-        return $this->stack()->toArray();
+        return $this->stack->toArray();
+    }
+
+    /**
+     * @return Collection
+     */
+    public function collect()
+    {
+        return clone $this->stack;
     }
 
     /**
@@ -47,7 +56,7 @@ trait StandardIteratorTrait
             throw new OutOfBoundsException('No items.');
         }
 
-        return $this->stack()->first();
+        return $this->stack->first();
     }
 
     /**
@@ -61,49 +70,49 @@ trait StandardIteratorTrait
             throw new OutOfBoundsException('No items.');
         }
 
-        return $this->stack()->last();
+        return $this->stack->last();
     }
 
     /**
-     * @param callable $callback
-     * @return StandardIterator
+     * @param Closure $callback
+     * @return StandardIteratorInterface
      */
-    public function filter(callable $callback)
+    public function filter(Closure $callback)
     {
         $filtered = new static();
-        $filtered->stack = $this->stack()->filter($callback);
+        $filtered->stack = $this->stack->filter($callback);
 
         return $filtered;
     }
 
     /**
-     * @param callable $callback
-     * @return StandardIterator
+     * @param Closure $callback
+     * @return StandardIteratorInterface
      */
-    public function reject(callable $callback)
+    public function reject(Closure $callback)
     {
         $filtered = new static();
-        $filtered->stack = $this->stack()->reject($callback);
+        $filtered->stack = $this->stack->reject($callback);
 
         return $filtered;
     }
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return bool
      */
-    public function every(callable $callback)
+    public function every(Closure $callback)
     {
-        return $this->stack()->every($callback);
+        return $this->stack->every($callback);
     }
 
     /**
-     * @param callable $callback
+     * @param Closure $callback
      * @return bool
      */
-    public function any(callable $callback)
+    public function any(Closure $callback)
     {
-        return $this->stack()->any($callback);
+        return $this->stack->any($callback);
     }
 
     /**
@@ -111,7 +120,7 @@ trait StandardIteratorTrait
      */
     public function isEmpty()
     {
-        return $this->stack()->isEmpty();
+        return $this->stack->isEmpty();
     }
 
     /**
@@ -119,7 +128,7 @@ trait StandardIteratorTrait
      */
     public function getIterator()
     {
-        return $this->stack()->copy();
+        return $this->stack->copy();
     }
 
     /**
@@ -127,18 +136,7 @@ trait StandardIteratorTrait
      */
     public function count()
     {
-        return $this->stack()->count();
+        return $this->stack->count();
     }
 
-    /**
-     * @return Collection
-     */
-    protected function stack()
-    {
-        if (!$this->stack instanceof Collection) {
-            $this->stack = new Collection();
-        }
-
-        return $this->stack;
-    }
 }